[Code][BP] Dynamic Rarity Algorithm
- Created a rarity selection algorithm based on level number.
- Filled all rarity pools.

Closes #48

+++ Source/Arena/ArenaGameMode.h
@@ -116,7 +116,7 @@ private:
 	UPROPERTY(BlueprintReadOnly, meta = (AllowPrivateAccess = "true"))
 	TArray<FCardData> CardsData;
 
-	int LevelNumber = 1;
+	uint8 LevelNumber = 1;
 	bool AllYellowsEnabled = false;
 
 	/*----------------------------------------------------------------------------
